Implemented infinite scrolling feed component (#62)
* Separated post types into multiple tables, and pushed migration

* Fixed errors/warnings from incorrect SVG property names

* Refactored rate limiters

* Added missing  conditions to schema, and pushed migration

* Refactored post, replies, and reposts endpoints into their own routers

* Added helper types to renamed  file

* Fixed typo

* Added missing router definitions to

* Implemented infinite scrolling feed component

* Utilizing edge runtime for home page

+++ src/server/api/routers/post.ts
@@ -1,64 +1,16 @@
 import { TRPCError } from "@trpc/server";
 import { and, eq } from "drizzle-orm";
 import { z } from "zod";
-import {
-  createTRPCRouter,
-  privateProcedure,
-  publicProcedure,
-} from "~/server/api/trpc";
+import { createTRPCRouter, privateProcedure } from "~/server/api/trpc";
 import { posts } from "~/server/db/schema";
-import { attachAuthors } from "~/utils/data";
 import { generateNanoId } from "~/utils/nanoid";
-import { postsRatelimiter } from "~/utils/ratelimiters";
+import ratelimit from "~/utils/ratelimit";
 
 export const postRouter = createTRPCRouter({
 
-  //
-  // PUBLIC PROCEDURES
-  //
-
-  getAll: publicProcedure.query(async ({ ctx }) => {
-    return await ctx.db.query.posts
-      .findMany({
-        orderBy: (posts, { desc }) => [desc(posts.createdAt)],
-        limit: 25,
-      }).then(attachAuthors);
-  }),
-
-  getPostsAndReposts: publicProcedure
-    .input(z.object({
-      userId: z.string().optional(),
-    }))
-    .query(async ({ ctx, input }) => {
-      return await ctx.db.query.posts
-        .findMany({
-          where: (posts, { and, eq, or }) => input.userId
-            ? and( eq(posts.userId, input.userId), or( eq(posts.type, "post"), eq(posts.type, "repost") ) )
-            : or(eq(posts.type, "post"), eq(posts.type, "repost")),
-          orderBy: (posts, { desc }) => [desc(posts.createdAt)],
-          limit: 25,
-        }).then(attachAuthors);
-    }),
-
-  getReplies: publicProcedure
-    .input(z.object({
-      userId: z.string().optional(),
-    }))
-    .query(async ({ ctx, input }) => {
-      return await ctx.db.query.posts
-        .findMany({
-          where: (posts, { and, eq }) => input.userId
-            ? and( eq(posts.userId, input.userId), eq(posts.type, "reply") )
-            : eq(posts.type, "reply"),
-          orderBy: (posts, { desc }) => [desc(posts.createdAt)],
-          limit: 25,
-        }).then(attachAuthors);
-    }),
-
-  //
-  // PRIVATE PROCEDURES
-  //
-
+  /**
+   * Creates a new post associated with the current authenticated user.
+   */
   create: privateProcedure
     .input(z.object({
       content: z.string().trim().min(1).max(280),
@@ -66,10 +18,10 @@ export const postRouter = createTRPCRouter({
     .mutation(async ({ ctx, input }) => {
       const userId = ctx.session.userId;
 
-      const { success } = await postsRatelimiter.limit(userId);
+      const { success } = await ratelimit.posts.create.limit(userId);
       if (!success) throw new TRPCError({
         code: "TOO_MANY_REQUESTS",
-        message: "You've reached your posting limit for the day!",
+        message: "You've reached your post creation limit for the day!",
       });
 
       await ctx.db.insert(posts).values({
@@ -79,14 +31,25 @@ export const postRouter = createTRPCRouter({
       });
     }),
 
+  /**
+   * Deletes a specified post associated with the current authenticated user.
+   */
   delete: privateProcedure
     .input(z.object({
       id: z.string().length(12),
     }))
     .mutation(async ({ ctx, input }) => {
+      const userId = ctx.session.userId;
+
+      const { success } = await ratelimit.posts.delete.limit(userId);
+      if (!success) throw new TRPCError({
+        code: "TOO_MANY_REQUESTS",
+        message: "You've reached your post deleting limit for the day!",
+      });
+
       await ctx.db.delete(posts).where(and(
         eq(posts.id, input.id),
-        eq(posts.userId, ctx.session.userId),
+        eq(posts.userId, userId),
       ));
     }),
